urlrig/requirements.txt
flask
flask-migrate 
flask-sqlalchemy 
psycopg2-binary
gunicorn 
python-decouple 
pyshorteners